all: Many changes, trying to move towards templated components.

+++ watcher.sh
@@ -2,8 +2,14 @@
 
 while true; do
     go build
-    CORRUGATION_AUTHENTICATION=false ./corrugation-backend &
-    PID="${!}"
-    inotifywait -r -e modify --include '.*\.go' .
-    kill "${PID}"
+    err="${?}"
+    if [ "${err}" != '0' ]; then
+        echo 'Failed to build...'
+        inotifywait -q -r -e modify --include '.*\.go' .
+    else
+        CORRUGATION_AUTHENTICATION=false ./corrugation-backend &
+        PID="${!}"
+        inotifywait -q -r -e modify --include '.*\.go' .
+        kill "${PID}"
+    fi
 done
